timer test + organizing
