+++ option.c
@@ -69,6 +69,7 @@ void set_full_screen(SDL_Window* window, Uint32 flags){ //FullScreenFlags
     }
 }
 
+/*
 void set_resolution(screen_state* screen,int w, int h){
     if(screen->fullscreen > 0){ //fullscreen
         //SDL_SetWindowDisplayMode();
@@ -78,3 +79,4 @@ void set_resolution(screen_state* screen,int w, int h){
     }
     
 }
+*/
